feat: self-contained agents damage calculation
